Move resource repository to a shareable location.
Anticpating use of ResolveResources by both the resource and application facades.
It could be moved to apiserver common, however there is already a charm related
repository in core/charm. Avoiding have 2 different things called a charm repository.
