+++ apiserver/facades/client/resources/package_test.go
@@ -6,29 +6,12 @@ package resources
 import (
 	"testing"
 
-	"github.com/juju/charm/v10"
-	charmresource "github.com/juju/charm/v10/resource"
 	gc "gopkg.in/check.v1"
-
-	corecharm "github.com/juju/juju/core/charm"
 )
 
 func Test(t *testing.T) {
 	gc.TestingT(t)
 }
 
-//go:generate go run github.com/golang/mock/mockgen -package mocks -destination mocks/charmhub.go github.com/juju/juju/apiserver/facades/client/resources CharmHub
 //go:generate go run github.com/golang/mock/mockgen -package mocks -destination mocks/logger.go github.com/juju/juju/apiserver/facades/client/resources Logger
 //go:generate go run github.com/golang/mock/mockgen -package mocks -destination mocks/backend.go github.com/juju/juju/apiserver/facades/client/resources Backend,NewCharmRepository
-
-func NewResourceClient(client CharmHub, logger Logger) *CharmHubClient {
-	c := &CharmHubClient{
-		Client: client,
-		logger: logger,
-	}
-	return c
-}
-
-func ResourceInfo(client *CharmHubClient, curl *charm.URL, origin corecharm.Origin, name string, revision int) (charmresource.Resource, error) {
-	return client.resourceInfo(curl, origin, name, revision)
-}
