+++ core/charm/repository/charmhub_test.go
@@ -7,12 +7,15 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"time"
 
 	"github.com/golang/mock/gomock"
 	"github.com/juju/charm/v10"
+	charmresource "github.com/juju/charm/v10/resource"
 	"github.com/juju/collections/set"
 	"github.com/juju/testing"
 	jc "github.com/juju/testing/checkers"
+	"github.com/juju/utils/v3/hash"
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/charmhub"
@@ -82,8 +85,7 @@ func (s *charmHubRepositorySuite) testResolve(c *gc.C, id string) {
 		origin.InstanceKey = "instance-key"
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	obtainedCurl, obtainedOrigin, obtainedSeries, err := repo.ResolveWithPreferredChannel(curl, origin)
+	obtainedCurl, obtainedOrigin, obtainedSeries, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, jc.ErrorIsNil)
 
 	curl.Revision = rev
@@ -119,8 +121,7 @@ func (s *charmHubRepositorySuite) TestResolveWithChannel(c *gc.C) {
 		},
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	obtainedCurl, obtainedOrigin, obtainedSeries, err := repo.ResolveWithPreferredChannel(curl, origin)
+	obtainedCurl, obtainedOrigin, obtainedSeries, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, jc.ErrorIsNil)
 
 	curl.Revision = 16
@@ -154,8 +155,7 @@ func (s *charmHubRepositorySuite) TestResolveWithoutBase(c *gc.C) {
 		},
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	obtainedCurl, obtainedOrigin, obtainedSeries, err := repo.ResolveWithPreferredChannel(curl, origin)
+	obtainedCurl, obtainedOrigin, obtainedSeries, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, jc.ErrorIsNil)
 
 	curl.Revision = 16
@@ -187,8 +187,7 @@ func (s *charmHubRepositorySuite) TestResolveWithBundles(c *gc.C) {
 		},
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	obtainedCurl, obtainedOrigin, obtainedSeries, err := repo.ResolveWithPreferredChannel(curl, origin)
+	obtainedCurl, obtainedOrigin, obtainedSeries, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, jc.ErrorIsNil)
 
 	curl.Revision = 17
@@ -221,8 +220,7 @@ func (s *charmHubRepositorySuite) TestResolveInvalidPlatformError(c *gc.C) {
 		},
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	obtainedCurl, obtainedOrigin, obtainedSeries, err := repo.ResolveWithPreferredChannel(curl, origin)
+	obtainedCurl, obtainedOrigin, obtainedSeries, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, jc.ErrorIsNil)
 
 	curl.Revision = 16
@@ -256,8 +254,7 @@ func (s *charmHubRepositorySuite) TestResolveRevisionNotFoundErrorWithNoSeries(c
 		},
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	_, _, _, err := repo.ResolveWithPreferredChannel(curl, origin)
+	_, _, _, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, gc.ErrorMatches,
 		`(?m)selecting releases: charm or bundle not found for channel "", platform "amd64"
 available releases are:
@@ -279,8 +276,7 @@ func (s *charmHubRepositorySuite) TestResolveRevisionNotFoundError(c *gc.C) {
 		},
 	}
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-	obtainedCurl, obtainedOrigin, obtainedSeries, err := repo.ResolveWithPreferredChannel(curl, origin)
+	obtainedCurl, obtainedOrigin, obtainedSeries, err := s.newClient().ResolveWithPreferredChannel(curl, origin)
 	c.Assert(err, jc.ErrorIsNil)
 
 	curl.Revision = 16
@@ -340,9 +336,7 @@ func (s *charmHubRepositorySuite) TestDownloadCharm(c *gc.C) {
 	s.expectCharmRefreshInstallOneFromChannel(c)
 	s.client.EXPECT().DownloadAndRead(context.TODO(), resolvedURL, "/tmp/foo").Return(resolvedArchive, nil)
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-
-	gotArchive, gotOrigin, err := repo.DownloadCharm(curl, requestedOrigin, "/tmp/foo")
+	gotArchive, gotOrigin, err := s.newClient().DownloadCharm(curl, requestedOrigin, "/tmp/foo")
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(gotArchive, gc.Equals, resolvedArchive) // note: we are using gc.Equals to check the pointers here.
 	c.Assert(gotOrigin, gc.DeepEquals, resolvedOrigin)
@@ -384,9 +378,7 @@ func (s *charmHubRepositorySuite) TestGetDownloadURL(c *gc.C) {
 
 	s.expectCharmRefreshInstallOneFromChannel(c)
 
-	repo := NewCharmHubRepository(s.logger, s.client)
-
-	gotURL, gotOrigin, err := repo.GetDownloadURL(curl, requestedOrigin)
+	gotURL, gotOrigin, err := s.newClient().GetDownloadURL(curl, requestedOrigin)
 	c.Assert(err, jc.ErrorIsNil)
 	c.Assert(gotURL, gc.DeepEquals, resolvedURL)
 	c.Assert(gotOrigin, gc.DeepEquals, resolvedOrigin)
@@ -411,9 +403,8 @@ func (s *charmHubRepositorySuite) TestGetEssentialMetadata(c *gc.C) {
 
 	s.expectCharmRefreshInstallOneFromChannel(c) // resolve the origin
 	s.expectCharmRefreshInstallOneFromChannel(c) // refresh and get metadata
-	repo := NewCharmHubRepository(s.logger, s.client)
 
-	got, err := repo.GetEssentialMetadata(corecharm.MetadataRequest{
+	got, err := s.newClient().GetEssentialMetadata(corecharm.MetadataRequest{
 		CharmURL: curl,
 		Origin:   requestedOrigin,
 	})
@@ -425,6 +416,153 @@ func (s *charmHubRepositorySuite) TestGetEssentialMetadata(c *gc.C) {
 	c.Assert(got[0].Manifest.Bases, gc.HasLen, 1)
 	c.Assert(got[0].ResolvedOrigin.ID, gc.Equals, "", gc.Commentf("ID is only added after charm download"))
 	c.Assert(got[0].ResolvedOrigin.Hash, gc.Equals, "", gc.Commentf("Hash is only added after charm download"))
+}
+
+func (s *charmHubRepositorySuite) TestResolveResources(c *gc.C) {
+	defer s.setupMocks(c).Finish()
+	s.expectRefresh(true)
+	s.expectListResourceRevisions(2)
+
+	result, err := s.newClient().ResolveResources([]charmresource.Resource{{
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginUpload,
+		Revision:    1,
+		Fingerprint: fp(c),
+		Size:        0,
+	}, {
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginStore,
+		Revision:    2,
+		Fingerprint: fp(c),
+		Size:        0,
+	}}, charmID())
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(result, gc.DeepEquals, []charmresource.Resource{{
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginUpload,
+		Revision:    1,
+		Fingerprint: fp(c),
+		Size:        0,
+	}, {
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginStore,
+		Revision:    2,
+		Fingerprint: fp(c),
+		Size:        0,
+	}})
+}
+
+func (s *charmHubRepositorySuite) TestResolveResourcesFromStore(c *gc.C) {
+	defer s.setupMocks(c).Finish()
+	s.expectRefresh(false)
+	s.expectListResourceRevisions(1)
+
+	id := charmID()
+	id.Origin.ID = ""
+	result, err := s.newClient().ResolveResources([]charmresource.Resource{{
+		Meta:     charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:   charmresource.OriginStore,
+		Revision: 1,
+		Size:     0,
+	}}, id)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(result, gc.DeepEquals, []charmresource.Resource{{
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginStore,
+		Revision:    1,
+		Fingerprint: fp(c),
+		Size:        0,
+	}})
+}
+
+func (s *charmHubRepositorySuite) TestResolveResourcesFromStoreNoRevision(c *gc.C) {
+	defer s.setupMocks(c).Finish()
+	s.expectRefreshWithRevision(1, true)
+
+	result, err := s.newClient().ResolveResources([]charmresource.Resource{{
+		Meta:     charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:   charmresource.OriginStore,
+		Revision: -1,
+		Size:     0,
+	}}, charmID())
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(result, gc.DeepEquals, []charmresource.Resource{{
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginStore,
+		Revision:    1,
+		Fingerprint: fp(c),
+		Size:        0,
+	}})
+}
+
+func (s *charmHubRepositorySuite) TestResolveResourcesNoMatchingRevision(c *gc.C) {
+	defer s.setupMocks(c).Finish()
+	s.expectRefresh(true)
+	s.expectRefreshWithRevision(99, true)
+	s.expectListResourceRevisions(3)
+
+	_, err := s.newClient().ResolveResources([]charmresource.Resource{{
+		Meta:     charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:   charmresource.OriginStore,
+		Revision: 1,
+		Size:     0,
+	}}, charmID())
+	c.Assert(err, gc.ErrorMatches, `charm resource "wal-e" at revision 1 not found`)
+}
+
+func (s *charmHubRepositorySuite) TestResolveResourcesUpload(c *gc.C) {
+	defer s.setupMocks(c).Finish()
+	s.expectRefresh(false)
+
+	id := charmID()
+	id.Origin.ID = ""
+	result, err := s.newClient().ResolveResources([]charmresource.Resource{{
+		Meta:     charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:   charmresource.OriginUpload,
+		Revision: 3,
+		Fingerprint: charmresource.Fingerprint{
+			Fingerprint: hash.Fingerprint{}},
+		Size: 0,
+	}}, id)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(result, gc.DeepEquals, []charmresource.Resource{{
+		Meta:     charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:   charmresource.OriginUpload,
+		Revision: 3,
+		Fingerprint: charmresource.Fingerprint{
+			Fingerprint: hash.Fingerprint{}},
+		Size: 0,
+	}})
+}
+
+func (s *charmHubRepositorySuite) TestResourceInfo(c *gc.C) {
+	defer s.setupMocks(c).Finish()
+	s.expectRefreshWithRevision(25, false)
+
+	curl := charm.MustParseURL("ch:amd64/focal/ubuntu-19")
+	rev := curl.Revision
+	channel := corecharm.MustParseChannel("stable")
+	origin := corecharm.Origin{
+		Source:   corecharm.CharmHub,
+		Type:     "charm",
+		Revision: &rev,
+		Channel:  &channel,
+		Platform: corecharm.Platform{
+			OS:           "ubuntu",
+			Channel:      "20.04",
+			Architecture: "amd64",
+		},
+	}
+
+	result, err := s.newClient().resourceInfo(curl, origin, "wal-e", 25)
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(result, gc.DeepEquals, charmresource.Resource{
+		Meta:        charmresource.Meta{Name: "wal-e", Type: 1, Path: "wal-e.snap", Description: "WAL-E Snap Package"},
+		Origin:      charmresource.OriginStore,
+		Revision:    25,
+		Fingerprint: fp(c),
+		Size:        0,
+	})
 }
 
 func (s *charmHubRepositorySuite) expectCharmRefreshInstallOneFromChannel(c *gc.C) {
@@ -555,6 +693,114 @@ func (s *charmHubRepositorySuite) setupMocks(c *gc.C) *gomock.Controller {
 
 func (s *charmHubRepositorySuite) expectedCURL(curl *charm.URL, revision int, arch string, series string) *charm.URL {
 	return curl.WithRevision(revision).WithArchitecture(arch).WithSeries(series)
+}
+
+func (s *charmHubRepositorySuite) newClient() *CharmHubRepository {
+	return NewCharmHubRepository(s.logger, s.client)
+}
+
+func (s *charmHubRepositorySuite) expectRefresh(id bool) {
+	s.expectRefreshWithRevision(0, id)
+}
+
+func (s *charmHubRepositorySuite) expectRefreshWithRevision(rev int, id bool) {
+	resp := []transport.RefreshResponse{
+		{
+			Entity: transport.RefreshEntity{
+				CreatedAt: time.Date(2020, 7, 7, 9, 39, 44, 132000000, time.UTC),
+				Download:  transport.Download{HashSHA256: "c97e1efc5367d2fdcfdf29f4a2243b13765cc9cbdfad19627a29ac903c01ae63", Size: 5487460, URL: "https://api.staging.charmhub.io/api/v1/charms/download/jmeJLrjWpJX9OglKSeUHCwgyaCNuoQjD_208.charm"},
+				ID:        "jmeJLrjWpJX9OglKSeUHCwgyaCNuoQjD",
+				Name:      "ubuntu",
+				Resources: []transport.ResourceRevision{
+					resourceRevision(rev),
+				},
+				Revision: 19,
+				Summary:  "PostgreSQL object-relational SQL database (supported version)",
+				Version:  "208",
+			},
+			EffectiveChannel: "latest/stable",
+			Error:            (*transport.APIError)(nil),
+			Name:             "postgresql",
+			Result:           "download",
+		},
+	}
+	s.client.EXPECT().Refresh(gomock.Any(), charmhubConfigMatcher{id: id}).Return(resp, nil)
+}
+
+// charmhubConfigMatcher matches only the charm IDs and revisions of a
+// charmhub.RefreshMany config.
+type charmhubConfigMatcher struct {
+	id bool
+}
+
+func (m charmhubConfigMatcher) Matches(x interface{}) bool {
+	config, ok := x.(charmhub.RefreshConfig)
+	if !ok {
+		return false
+	}
+	h, err := config.Build()
+	if err != nil {
+		return false
+	}
+	if m.id && h.Actions[0].ID != nil && *h.Actions[0].ID == "meshuggah" {
+		return true
+	}
+	if !m.id && h.Actions[0].Name != nil && *h.Actions[0].Name == "ubuntu" {
+		return true
+	}
+	return false
+}
+
+func (m charmhubConfigMatcher) String() string {
+	if m.id {
+		return "match id"
+	}
+	return "match name"
+}
+
+func (s *charmHubRepositorySuite) expectListResourceRevisions(rev int) {
+	resp := []transport.ResourceRevision{
+		resourceRevision(rev),
+	}
+	s.client.EXPECT().ListResourceRevisions(gomock.Any(), gomock.Any(), gomock.Any()).Return(resp, nil)
+}
+
+func fp(c *gc.C) charmresource.Fingerprint {
+	fp, err := charmresource.ParseFingerprint("38b060a751ac96384cd9327eb1b1e36a21fdb71114be07434c0cc7bf63f6e1da274edebfe76f65fbd51ad2f14898b95b")
+	c.Assert(err, jc.ErrorIsNil)
+	return fp
+}
+
+func charmID() corecharm.CharmID {
+	curl := charm.MustParseURL("ubuntu")
+	channel, _ := charm.ParseChannel("stable")
+	return corecharm.CharmID{
+		URL: curl,
+		Origin: corecharm.Origin{
+			ID:      "meshuggah",
+			Source:  corecharm.CharmHub,
+			Channel: &channel,
+			Platform: corecharm.Platform{
+				OS:           "ubuntu",
+				Channel:      "20.04",
+				Architecture: "amd64",
+			},
+		}}
+}
+
+func resourceRevision(rev int) transport.ResourceRevision {
+	return transport.ResourceRevision{
+		Download: transport.Download{
+			HashSHA384: "38b060a751ac96384cd9327eb1b1e36a21fdb71114be07434c0cc7bf63f6e1da274edebfe76f65fbd51ad2f14898b95b",
+			Size:       0,
+			URL:        "https://api.staging.charmhub.io/api/v1/resources/download/charm_jmeJLrjWpJX9OglKSeUHCwgyaCNuoQjD.wal-e_0",
+		},
+		Name:        "wal-e",
+		Revision:    rev,
+		Type:        "file",
+		Filename:    "wal-e.snap",
+		Description: "WAL-E Snap Package",
+	}
 }
 
 // RefreshConfigMatcher is required so that we do check somethings going into
